requirements.txt
attrs==24.3.0
autopep8==2.3.1
certifi==2024.12.14
h11==0.14.0
idna==3.10
numpy==2.2.1
outcome==1.3.0.post0
pycodestyle==2.12.1
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
